db: Add priority mix feature in db and object
Add string column "priority" to the instances table in DB and
corresponding field in the Instance versioned object. Add column
"priority", "high_vcpus_used" and "low_vcpus_used" to the
compute_nodes table in DB and corresponding field in the versioned
object.

The "priority" is used to identify the high or low priority of the
instances. If priority is not set, the current instance is a common
instance with no priority.